Add shuffleData and cardOnClick methods to App
Implemented shuffleData and cardOnClick methods in the App component to
enable shuffling of card data. shuffleData utilizes the Fisher-Yates
Shuffle algorithm to randomize the order of an array, while cardOnClick
updates the card data state with a shuffled version of the current card
data array when a card is clicked.

These methods enhance the difficulty of the card game.